Rename package from "constructors" to "models" & Create Priority class
