--- src/main/java/common/constructors/Priority.java
@@ -1,29 +1,0 @@
-package common.constructors;
-
-public class Priority {
-
-    String name;
-
-    int priority;
-
-    Priority(String name, int priority) {
-        this.name = name;
-        this.priority = priority;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getPriority() {
-        return priority;
-    }
-
-    public void setPriority(int priority) {
-        this.priority = priority;
-    }
-}
