+++ src/main/java/common/models/Member.java
@@ -1,4 +1,4 @@
-package common.constructors;
+package common.models;
 
 /**
  *
