+++ src/main/java/common/models/Restrictions.java
@@ -1,35 +1,35 @@
-package common.constructors;
+package common.models;
 
 public class Restrictions {
 
-    // Имеет ли возможность кикать
+    // Ограничение на использование команды кик
     public Limit canKick;
 
-    // Имеет ли возможность выдавать бан
+    // Ограничение на использование команды бан
     public Limit canBan;
 
-    // Имеет ли возможность снимать бан
+    // Ограничение на использование команды для разбана
     public Limit canUnban;
 
-    // Имеет ли возможность выдавать мут
+    // Ограничение на использование команды мут
     public Limit canMute;
 
-    // Имеет ли возможность снимать мут
+    // Ограничение на использование команды снятия мута
     public Limit canUnMute;
 
-    // Имеет ли возможность выдавать предупреждение
+    // Ограничение на использование команды выдачу предупреждения
     public Limit canWarn;
 
-    // Имеет ли возможность снимать предупреждение
+    // Ограничение на использование команды снятия предупреждения
     public Limit canRemWarn;
 
-    // Имеет ли возможность обнулять предупреждения
+    // Ограничение на использование команды сброса предупреждений
     public Limit canResetWarn;
 
-    // Имеет ли возможность очищать сообщение(-я) в чате
+    // Ограничение на использование команды очищать сообщение(-я) в чате
     public Limit canClear;
 
-    // Имеет ли возможность выдавать временную роль
+    // Ограничение на использование команды выдавать временную роль
     public Limit canTempRole;
